cleanup and fix fatal and panic
